+++ hnix-store-core/CHANGELOG.md
@@ -10,6 +10,7 @@
    * `BuildResult`s `timesBuild` field changes type from `Integer` to `Int` [#231](https://github.com/haskell-nix/hnix-store/pull/231)
 
 * Additions:
+   * `System.Nix.StorePath.parsePathFromText` [#236](https://github.com/haskell-nix/hnix-store/pull/236)
    * `Default StoreDir` instance [#231](https://github.com/haskell-nix/hnix-store/pull/231)
    * `System.Nix.StorePath.storePathHashPartToText` [#231](https://github.com/haskell-nix/hnix-store/pull/231)
    * Added `Generic` and `Show` instances for